# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cape Christian Sermon Planning API
  description: RESTful API for sermon planning and management system
  version: 1.0.0
  contact:
    name: Cape Christian Tech Team
    email: tech@capechristian.org
servers:
  - url: https://api.sermons.capechristian.org/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /series/{id}/sermons:
    get:
      tags: [Series]
      summary: Get sermons in series
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        200:
          description: List of sermons in series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sermon'

  # Sermons
  /sermons:
    get:
      tags: [Sermons]
      summary: List sermons
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: seriesId
          in: query
          schema:
            type: string
            format: uuid
        - name: speakerId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, review, approved, delivered, archived]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of sermons
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sermon'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Sermons]
      summary: Create new sermon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SermonCreate'
      responses:
        201:
          description: Sermon created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sermon'
  /sermons/{id}:
    get:
      tags: [Sermons]
      summary: Get sermon by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        200:
          description: Sermon details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sermon'

    put:
      tags: [Sermons]
      summary: Update sermon
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SermonUpdate'
      responses:
        200:
          description: Sermon updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sermon'

    delete:
      tags: [Sermons]
      summary: Delete sermon
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        204:
          description: Sermon deleted
  /sermons/{id}/timeline:
    get:
      tags: [Sermons]
      summary: Get sermon timeline
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        200:
          description: Timeline phases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelinePhase'

    post:
      tags: [Sermons]
      summary: Update timeline phase
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimelinePhaseUpdate'
      responses:
        200:
          description: Timeline updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelinePhase'

  # Timeline
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Sermon:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        sermonNumber:
          type: integer
        speakerId:
          type: string
          format: uuid
        speakerName:
          type: string
        scheduledDate:
          type: string
          format: date
        deliveredDate:
          type: string
          format: date
        serviceTime:
          type: string
          format: time
        durationMinutes:
          type: integer
        primaryScripture:
          type: string
        scriptureReferences:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, review, approved, delivered, archived]
        timelinePhase:
          type: string
        callToAction:
          type: string
        targetAudience:
          type: string
        ministryFocus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SermonCreate:
      type: object
      required: [title, scheduledDate]
      properties:
        seriesId:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        sermonNumber:
          type: integer
        speakerName:
          type: string
        scheduledDate:
          type: string
          format: date
        serviceTime:
          type: string
          format: time
        primaryScripture:
          type: string
        scriptureReferences:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        callToAction:
          type: string
        targetAudience:
          type: string
        ministryFocus:
          type: string
    SermonUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        sermonNumber:
          type: integer
        speakerName:
          type: string
        scheduledDate:
          type: string
          format: date
        deliveredDate:
          type: string
          format: date
        serviceTime:
          type: string
          format: time
        durationMinutes:
          type: integer
        primaryScripture:
          type: string
        scriptureReferences:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, review, approved, delivered, archived]
        timelinePhase:
          type: string
        callToAction:
          type: string
        targetAudience:
          type: string
        ministryFocus:
          type: string

    # Timeline Schemas
    TimelinePhase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sermonId:
          type: string
          format: uuid
        phase:
          type: string
          enum: [brainstorm, content_planning, wordsmith, review, production, rehearsal, delivery, archive]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        assignedTo:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        notes:
          type: string
        checklist:
          type: array
          items:
            type: object
            properties:
              task:
                type: string
              completed:
                type: boolean
    TimelinePhaseUpdate:
      type: object
      properties:
        phase:
          type: string
          enum: [brainstorm, content_planning, wordsmith, review, production, rehearsal, delivery, archive]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        assignedTo:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        notes:
          type: string
        checklist:
          type: array
          items:
            type: object
            properties:
              task:
                type: string
              completed:
                type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...